+++ tests/test_moti_download.py
@@ -2,7 +2,7 @@ import datetime
 
 import pytest
 
-import crmprtd.download
+import crmprtd.download_utils
 from crmprtd.moti.download import download
 
 
@@ -45,15 +45,15 @@ def test_download(stime, etime, station_id, expected_payload, mocker):
     # We don't need/want to test the actually https request going out
     # so let's just mock that call and ensure that it gets called with
     # the expected arguments
-    mocker.patch("crmprtd.download.https_download")
+    mocker.patch("crmprtd.download_utils.https_download")
     # Default time arguments are the "present" time at the time
     # download() is run. Since this is not predictable, we need to
     # mock out datetime.utcnow() to give us a deterministic time to
     # test against.
     mocker.patch("crmprtd.moti.download.utcnow", return_value=now)
     download("u", "p", None, None, stime, etime, station_id, url)
-    crmprtd.download.https_download.assert_called_once()
-    call_args, _ = crmprtd.download.https_download.call_args
+    crmprtd.download_utils.https_download.assert_called_once()
+    call_args, _ = crmprtd.download_utils.https_download.call_args
     _, _, _, _, payload = call_args
     assert payload == expected_payload
